Idiomatized MergeSort.hs
The implementation as it was suffered from "boolean blindness" which is a common occurence in translations of imperative algorithms into a functional style. It describes the phenomenon that structural properties of data and contextual knowledge from branches and pattern-matching is disregarded in favor of boolean-valued tests. E.g. the cascading branches on the length of the list in `mergeSort` can be much more succinctly expressed by a collection of patterns.

Quick notes for idiomatic Haskell:
- favor pattern-matching over conditionals
- favor pattern-matching over unsafe destructors (`head`, `tail`, etc).
- favor `where` blocks over `lets`